jotting_jwts/jotting_jwts/wsgi.py
"""
WSGI config for jotting_jwts project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/3.1/howto/deployment/wsgi/
"""

import os

from django.core.wsgi import get_wsgi_application
from dotenv import load_dotenv
load_dotenv()

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'jotting_jwts.settings')

application = get_wsgi_application()
